Finished UI health bar
UI health bar is finished. Once enemy(s) take damage health bar does go down and also disappear when player dies. May work on hit only activating when taking damage next

+++ Assets/Scripts/IDamage.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 	public interface IDamage
 	{
-		public void takeDamage(int dmg);
+		public void takeDamage(float dmg);
 	
 	}
